Setup speed and duplex with vendor specific goo under Win7.
There is no standard interface for setting NIC speed and duplex under windows.
Hack in some goo to use vendor tooling to do this.  Done for Broadcom.  Still
need to do Intel.  A change to assign may come along and allow us to remove
this stuff in a while (when Rob has time).  This vendor-specific stuff will
be a royal PITA for other sites.

A