Update OS command line on delay nodes if node is reloaded with a new OS.
Every once in a while, I change the version of FreeBSD running on a delay
node (with os_load) and it would blow up because the command line to boot
the delay kernel with would be wrong.

Note that this is not a general fix for default command lines, it is limited
to delay node kernels.