Add support for creating snapshot images of OpenVZ containers via the web interface.
Most of this commit is about making it easier to create the
descriptors (of any kind) The newimageid_ez page now takes an optional
node_id on which to base the new descriptor; most of the fields are
initialized.

You can also provide optional imageid and osinfo arguments, which
allows for slightly less info to be initialized.

Then on the ShowNode page, the link to create an image will take you
to the snapshot image page if the user has write access to the image,
otherwise to the newimageid_iz page to create and snapshot a new
image descriptor.