A variety of tweaks to make up/download of images on vnodes work.
In the master server, recognize "sharedhost" as a valid proxy node so
that shared vnode host nodes can upload and download images on behalf
of hosted vnodes that are not in the same experiment.

Add "-P <node-i-am-a-proxy-for>" options to frisbee and frisupload clients
allowing them to request images on behalf of other nodes. Previously,
proxying was only really supported in nested instances of the master server.

Make sure that such proxying works when a shared vnode host is serviced by
a subboss; i.e., subboss is proxying for the vnode host which in turn is
proxying for the vnode.