Fixes for Windows 7 base setup, phase 1.
'sc' is an alias for the "Set-Content" powershell cmdlet, which collided with
me attempting to invoke the sc system command.  Haven't found a way to prevent
alias expansion, so am just specifying the full path for now.  Also adjust the
way the clock (ntp) is configured and fix a bug in the registry modification
code.