More bits and pieces for exporting images from one Emulab to another.
image_metadata.php will return an Emulab style image descriptor in XML
format. A remote emulab, given an image URL, will grab this XML
description and use it to create a local descriptor. Inside the
descriptor is an additional URL that is used to download ndz file.

The dumpdescriptor script is now web accessible, and takes a new -e
(export) option that adds the extra URL and other bits that are needed
to import the descriptor and the image.

On the Show Image page, show the metadata URL, which is suitable for
using in an NS file or an rspec (when that code is committed).