Add SSL-only "arpinfo" command to return MAC/IP for servers and "peer" nodes.
Details:

Return MAC/IP (ARP) information for a node's "peers" on the node control
net. We also return info for the control net gateway (if there is one)
and any servers on the node control network (e.g., subbosses).

As far as "peer" info, right now we just support calls by subbosses that
will return the info for the set of nodes they control. This could also
be used by experiment firewalls or gateways to return info for the nodes
they are protecting and setup filters accordingly.

Note that this call only returns info if invoked using SSL. This allows
the client to be sure it is getting the info from boss.

Tangentially related, is the addition of an index to the subbosses table
to speed up a really slow query.