"improvements" to prereserve:
New option -s datetime to specify a starting time for the pre-reserve.
New option -e datetime to specify a ending time for the pre-reserve.

The idea is that you can schedule a pre-reserve to begin sometime later,
and you can optionally specify a time for a prereserve to terminate.
There is a new script that runs from cron that checks for pre-reserves
that need to be started or terminated.

For example:

boss> wap prereserve -s '2012-09-14 09:08:15' -e '2012-09-15' emulab-ops 50

You can use any datetime string that is valid for str2time. At some point
it would be nice to allow natural language dates ("tomorrow") but that
requires a another bunch of perl packages and I didn't want to bother.

NOTE: when using -e, -r is implied; in other words, when the
pre-reserve is terminated, the table entry is cleared *and* the
reserved_pid of all of the nodes is cleared. Any experiments using
those nodes is left alone, although if the user does a swapmod, they
could easily lose the nodes if another pre-reserve is set up that
promises those nodes to another project.