Small fixes and improvements.
Added ability to force AM API for managers, help button on the login page since that seems to be a source of confusion, live updating of the slice/slivers expires time so it reflects the actual time left instead of how much time was left the last time it was checked, global checkbox for creating private or global images.

AM API v3 support has started. So far the client is able to list resources at a manager. More work will need to be done to support the new create workflow now that it supports a two-phased commit similar to the CM API.

Fixed user being able to create a link while the slice is locked (having calls made), crash when user comes in with a credential rather than using a slice authority, and new slice button being shown even if user can't create slices.