Brutalize to use ipfw command line interface when > FreeBSD 8.
The socket interface changed and rather rewrite this client or try to
figure out how to finish the "new" client, I whacked the crap out of
this to use ipfw. This is of course more fragile since it depends on
the input language of ipfw, and the especially the output from ipfw to
determine the current settings. When upgrading, we must be sure to
test this.