Add the -x "exclude non-FS" option to imagezip.
This option says to ignore any blocks which are outside of a Linux or FreeBSD
filesystem but still in the containing partition; i.e., in the case where the
filesystem is smaller than the partition it is in. By default, such space is
(and always has been) considered "allocated."

This should be extended to cover space that is outside of any FreeBSD or
DOS partition as well (though we have to be careful not to exclude the MBR
in that case). It should also be made to work with FAT and NTFS.