Move rpm/tar download from boss to ops, to avoid wasted network traffic.
To turn this option on, define SPEWFROMOPS=1 in your defs file. This
will result in a redirect message from boss which will send the wget
client over to ops. 

A perl setuid root cgi script is run from the webserver on ops when a
/spewrpmtar request is made. This script sends the key,nodeid,file
over to boss via XMLRPC (as elabman). The return is simple yes or no,
the caller is allowed (not allowed) to have that file. Since the
ops script runs as root, it can spew the file back to the caller.

Note that the elabinelab checks for the elabinelab source code are
gone; we are now open source. Also, we spew that file from /share now,
to be consistent.