More PRObE inspired improvements to the swapin path.
Replace an exec of the perl os_select script with a call to the OSSelect()
node method. This cut in half the time spent in the DB setup for each node.
Note that this change had already been made to libosload_new.

Reworked the code that setup the partitions table entries. We were potentially
updating each DB row for each image loaded. Now we just work out all the changes
in a perl data struct and make one set of DB changes at the end. The code is
more comprehensible now as well (I hope!)

Finally, disable the "swapinfo" stuff which was the first step in doing stateful
swapout of disk state. That code never got finished.