More ARP lockdown related changes.
Make sure sitevars get initialized on initial installation of an Emulab.
Fixes to the update_sitevars script, mostly in case we someday want to
run it on every testbed software install (which we do not do right now).

For ops and fs there is a race with boss that prevents us from locking
down ARP entries early. For now, we do the lock down later in the boot.
If someone spoofs boss or the gateway before then, we will detect it
when we request the ARP info via SSL-enabled tmcc.