Start slothd inside openvz containers.
Change the logging code so that we get better output on early failure.
Openvz nodes do not have a "console" so early errors are lost. This is
a real pain, cause before this point, we have no idea what went wrong.
Later versions of openvz are supposed to have a console.