Initial commit for work on Flack 3
The work needed to support AM API v3/v4 was extensive enough that a major
rehaul was required to support the new paradigms like allocating and
provisioning as opposed to creating. A lot of preparation has been done
to fix various annoyances like going back from a failed sliver creation
to a stagged sliver.  Previously it was suggested to just delete the
entire slice, but it would be much better to be able to more easily
move between stagged/allocated/provisioned without having to start
from scratch.