Add creator_urn to images table so we can track actual creator when created via the ProtoGeni interface; we run everything as GeniUser and so lose that info. We want to know this for history information and so we can manage permissions on image deletion effectively.
Also add updater, updater_idx, and updater_urn to track last change
to the image.

CAVEAT: For system images, we do not actually overwrite the image, but
store it out in /proj/$pid/images and then sometime later maybe we
copy it back to boss. This wrecks the tracking info quite effectively
since the image may never be copied back. Needs more thought.