Add long term storage of log files in /usr/testbed/logs/logfiles.
New Store() method will gzip and move the file from current location
to long term storage, and update the DB row to reflect the change.
Change spewlogfile to handle these compressed files. Also added a
"public" flag that says a log file can be requested by an anonymous
user that knows the logid (md5 of bits). Change spewlogfile to allow
anonymous requests.

Add logfile_metadata to hold key,value pairs associated with a
logfile, eventually for search but initially to include when returning
the contents of a logfile via spewlogfile.