Various changes and fixes to libaudit.
Add TBLOGS option to send LOGONLY logs to testbed-logs instead of
default tbops (except when an error).

Do not delete logfiles when NODELETE is given; we were not being
careful about that.

Changes and fixes to AuditFork to support new usage from ProtoGeni
XMLRPC (the only place we use it).