Change calls to AuditFork() to main::WrapperFork(). We want to know
that the callee has forked in the wrapper so we can avoid storing the
log twice into long term storage. Originally, we just created a new
log file since it was going to be deleted anyway, but now we want to
just extend the original logfile for long term storage.