Lots of change to logfile handling.
Save logs to long term storage. See changes to db/Logfile.pm.in

Add WrapperFork() so the wrapper knows the callee has forked (the
log) so that we can extend the current log instead of creating a
second logfile.

Add all of the message header lines (except results/args) to new
logfile_metdata table.

Create and add logurn header so that the user can go find the log for
the operation.