Fix for "orphaned certificate" warnings we got this week.
So this happened twice this week. Two Utah Emulab users, starting a
sliver at the Utah Emulab CM. Basically, while Flack was starting up a
sliver for the user, they decided to log into the web interface and
recreate their encrypted certificates. So they register a slice, and
Flack gets the slice certificate. Then the user changes their
certificate on Emulab. The CM notices that the certificate in the
slice credential and the certificate the user presented are
different. For a nonlocal user we want to update our record (dubious
on its own), but for a local user we really do not want to do
anything.

Users do wacky things.