Changes to image storage to prevent fill up the local disk.
All images are now stored in their own lvm and the tar file is deleted
once that is created. We no longer need storage on / or any other file
system, except for the initial download. For that we use a scratch FS
and delete the tar file right away.

A medium cleanup of the image and lvm handling, although mostly I
brought across the cleanups I did to the xen version of the same
code.

Add a new script called prunelvms that runs from cron on SHARED nodes
each night, that will remove the image lvms that are no longer in use
and have not been used for 24 hours.