Fix the problem with slow network bandwidth between VMs on the same node.
This problems appears when creating a lan of VMs on the same physical
host, where the macvlan device is on top of a real ethernet interface.
The BW was on the order of a 100s of Kbits, which is terrible. But
crossing to a node on another physical host was fine. I eventually
stumbled onto the fact that if the underlying device has tso/gso on,
bw sucked. Just turning that off solved the problem. I have no idea
why, just accept. 

Also turn off ipv6; some kind of unrelated kernel leak that was
causing messages to the console, and I hate those while I type.