Add a check to prevent a link or lan between a VM and a physical node, if the OS on the physical node does not support linkdelays.  Most OS do not, and it is wrong to allow this to happen since traffic at the physical host will not be subject to shaping. I doubt this happens very often, but it was on a Geni acceptance test, wich was showing asymetric behavour when testing bw.
Giant Caveat: This will do the right thing on the Geni racks, but on
Emulab and perhaps other clusters it might not. That is because Utah
has some node types with both 100Mb and 1Gb interfaces, and as far are
the mapper is concerned, these links do not need to be shaped if you
ask for a 100Mb link. So we do not do the check and the BW will not be
correct. I think there is a way to deal with this, but I am not sure
yet, and I need to get this fix in so it can go to the racks.