Changes to get GRE tunnels into /etc/hosts when coming from PG.
The main problem is that when coming from the PG interface, we do not
create normal virt_lans entries for GRE tunnels, since one side of the
tunnel in not a local node and that would confuse all kinds of things.
So instead we enter that directly into the lans table when a ticket is
redeemed. So, gentopofile never sees them and they do not go into the
"topofile"

So, I've added a pass in gentopofile that checks the lans table for
tunnels that are not in virt_lans.