Reorder iSCSI teardown sequence and add a wait.
Reorder the teardown sequence walked through for undoing the "EXPORT" storage
directive on FreeNAS.  Hopefully this improves the reliability of this
procedure.  Also added a wait after the iSCSI extent teardown so that whatever
might still be holding the underlying ZVOL has time to let go before we try
to delete it.