Try to make getimages resilient to interruption or temporary failure.
The basic approach is to tell fetch to be optimistic (preserve partial
output, attempt to resume transfers when local data exist, etc.).  We
ignore all errors fetch gives us, and rely on the SHA1 hashes only for
integrity verification.  If we think our copy is too small, we just
keep trying to resume.  If our copy is big enough but the checksum
doesn't match, then throw it away and restart.