Support for two new "loadinfo" options BIOSDISK and CONSOLE.
BIOSDISK is passed through from the DB bootdisk_bios_id and is used to
set the GRUB root device (in situations where grub disagrees with freebsd
or linux about the primary boot device).

CONSOLE is used to set the console device for an image's boot loader and
kernel. Choices are "null", "vga", "sio", (the historic choices) along
with "vid" (same as vga), "sio1" (same as sio), "sio2", "sio3", and "sio4".
The FreeBSD slicefix script works for FreeBSD 9.x and above as well as
Fedora, Ubuntu and CentOS. It will work with FreeBSD 8.x and earlier if
the in-image /boot/loader is fixed to avoid hanging while polling a UART.

This should be used in conjuction with one of the new
pxeboot.emu-{null,vga,sio1,sio2,sio3,sio4} versions of pxeboot which
will customize the FreeBSD kernel in the frisbee and admin MFSes.

The console can be set via the "console_type" node_ or node_type_attribute
tables in the DB. The "pxe_boot_path" attribute should also be set to
reflect the appropriate /tftpboot/pxeboot.emu-* version for the console type

Note that by default (no console_type or pxe_boot_path), nodes will continue
to use the default /tftpboot/pxeboot.emu which can either be linked to one
of the above new version, or could just be the current pxeboot which will
not mess with the console settings.

Finally, note that I made the changes to linux/slicefix, but they are only
for FreeBSD images and are untested. The Linux grub-based pxeboot as well
as linux_slicefix would still need to be changed for this to work under the
"Linux MFS". Task for another day as I already spent waaaaay to long on this.