Make it explicit that dhcpd is not restarted on an error.
Simply ignoring the error and letting dhcpd restart doesn't work
because dhcpd_makeconf doesn't actually ever restart dhcpd, rather
it relies on the dhcpd_wrapper. But on an error, the wrapper has
died as well. So now the error message just says that dhcpd is NOT
running and to use the startup script to restart things.