Various bug fixes and initial stitching implementation
There are a lot of files renamed to remove the keyword 'virtual' from their
path. Apparently something recently changed which makes Flash Builder not like
paths with 'virtual'.

ExoGENI racks are now supported. One extra thing they required was the ability
to edit disk image versions. That's a required value for them so I had to expose
it in the node properties editor.

The XmlRpc parser choked on parsing data values which didn't specify what type
they are.  The spec says to parse those nodes as string values so that's what
the library does now.

Some users noted that Flack, when embedded in an external page, didn't load all
managers correctly initially. I got around that by making sure to try each
manager twice before concluding the manager is unreachable.  This doesn't fix
whatever the underlying problem is (since many still fail the first time) but it
does cause the managers to get loaded correctly.

The implementation of stitching worked in a simple test connecting a node from
Emulab to the Utah Instageni rack. Parsing of advertised and request stitching
data was completed and an extra step was added to SubmitSlice to do any
stitching before actually submitting the slice. Now if any stitching is done and
a workflow is received through that step, dependencies will cause the allocation
steps to occur in the correct order to satisfy the stitching constraints.