Fix suddenly-appearing bug in libosload_new dynamic module loading code.
Not sure why this change was needed, but it wasn't working anymore.
Maybe this is a perl 5.10 -> 5.12 thing; how I was doing an eval of a
require to load an libosload_X module was no longer working.  The new
way is better anyway, but even after looking through the perlXdelta
pages, I didn't see a reason for this.

Oh well.