Add complete local node storage support from parser down to tcmd.
Doing this required adding columns to the virt and physical blockstores
tables to mark the attributes that will be considered for mapping.
Unmarked entries just flow through to the client-side.

This commit also introduces filesystem support in the form of passing
through a mount point to the client-side.  It is left to the client to
decide what filesystem and fs options to use to setup the space, including
any logical volume aggregation required to support the request.