Add syntactic sugar for attaching a SAN storage object to node.
Instead of requiring users to specify a hokey-looking duplex link
between a remote blockstore and a node, let them use the "set-node"
method on the blockstore object, which implicitly creates this link.

Also, do the same for lans:  introduce [$ns make-san $nodelist]

If/when in the future we allow experimenters to shape links to storage
servers, specifying the link directly will make more sense.