Changes to support high vlans on individual wires.
Utah has rewired some of the its pc3000s to a procurve, which supports
vlans > 1024. Since we do not have link features and desires, we need
to trick assign by using a different link type on the wire, in both
the ptop and vtop. The vtop looks for a vlan reserved tag greater then
1024, and changes the linktype to ethernet-highvlan. In ptopgen we
look for an attribute on the switch that says it can do highvlans, and
add that linktype for any wire attached it.