Many bug fixes and add BW cap enforcement using IFB devices.
This replaces the previous revision that did this with an extra bridge
and IMQ device. We like IFBs better now! Note that we enforce just the
maximum BW; all other shaping is done in the current way, by mapping
an IMQ into the container and doing linkdelays as we do now.