Another round of improvements including ZFS and striping support.
Use ZFS rather than gvinum as the LVM on 64-bit FreeBSD. One funkiness:
if the blockstore has a mountpoint (implying a filesystem), we create a
ZFS filesystem. If they do not, we create a ZFS zvol (a block device).
The filesystem has much better performance than the zvol, hence using it
when we can.

In gvinum, if all disks are the same size, stripe them rather than
concanonate them. For us, this would only happen on the d820s if you
said NONSYSVOL (and you aren't using ZFS).

Support the unspecified (zero) length blockstore which means "as much
space as is available".

Handle mounting of SAN-based storage. We cannot just throw the mount
into /etc/fstab since the SAN device would not exist when fsck or
mount is done at boot time. You could probably configure it to work
on Linux, but I didn't see how on Ubuntu. The iSCSI initiator and
associated infrastructure on FreeBSD is so primitive that there is
no way you could do it automatically.

Add a "fullreset" command to rc.storage for debugging help. It attempts
to completely teardown the storage no matter what state it is in. The
regular "reset" command is conservative and quits at the first sign of
trouble.