Add OS features for local and remote blockstores.
loc-bstore: local blockstore support
rem-bstore: remote blockstore support

Check for these when a user attaches a blockstore to a local node (local
blockstores), or to a link/lan (remote blockstores).  All nodes in a lan
containing one or more remote blockstores must have an OS with the
remote blockstores feature.