Don't die if there is an error connecting to the icebox.
This would kill off the entire "power" invocation. For example, if the
first node in the list given to power failed to connect to its icebox,
power would die and not even attempt to power cycle any other node in
the list.

This manifested itself by lots of nodes getting stuck in reloading
because the reload_daemon would try to reboot every node in a single command.